added new path for solo and mutiplayer
