+++ socket_server/game/constants.js
@@ -1,0 +1,35 @@
+const TICK_RATE = 60;
+const TICK_MS = 1000 / TICK_RATE;
+const WORLD_HALF = 11;
+const PLAYER_RADIUS = 0.6;
+const PLAYER_MASS = 1;
+const MAX_PLAYERS_PER_ROOM = 8;
+const ACCEL = 22;
+const FRICTION = 0.88;
+const MAX_SPEED = 9;
+const RESTITUTION = 0.8;
+const TURN_SPEED = 3.4;
+const PLATE_RADIUS = WORLD_HALF;
+const GRAVITY = 24;
+const FALL_ELIMINATION_Y = -2;
+const SPAWN_DISTANCE = 6;
+const PLATE_SURFACE_Y = PLAYER_RADIUS;
+const PLATE_BOUNDARY_RADIUS = PLATE_RADIUS - PLAYER_RADIUS;
+
+module.exports = {
+  TICK_RATE,
+  TICK_MS,
+  PLAYER_RADIUS,
+  PLAYER_MASS,
+  MAX_PLAYERS_PER_ROOM,
+  ACCEL,
+  FRICTION,
+  MAX_SPEED,
+  RESTITUTION,
+  TURN_SPEED,
+  GRAVITY,
+  FALL_ELIMINATION_Y,
+  SPAWN_DISTANCE,
+  PLATE_SURFACE_Y,
+  PLATE_BOUNDARY_RADIUS,
+};
