+++ socket_server/game/room-service.js
@@ -1,188 +1,52 @@
-const TICK_RATE = 60;
-const TICK_MS = 1000 / TICK_RATE;
-const WORLD_HALF = 11;
-const PLAYER_RADIUS = 0.6;
-const PLAYER_MASS = 1;
-const MAX_PLAYERS_PER_ROOM = 8;
-const ACCEL = 22;
-const FRICTION = 0.88;
-const MAX_SPEED = 9;
-const RESTITUTION = 0.8;
-const TURN_SPEED = 3.4;
-const PLATE_RADIUS = WORLD_HALF;
-const GRAVITY = 24;
-const FALL_ELIMINATION_Y = -2;
-const SPAWN_DISTANCE = 6;
-const PLATE_SURFACE_Y = PLAYER_RADIUS;
-const PLATE_BOUNDARY_RADIUS = PLATE_RADIUS - PLAYER_RADIUS;
-
-function clamp(value, min, max) {
-  return Math.min(max, Math.max(min, value));
-}
-
-function randomRoomId() {
-  return `room_${Math.random().toString(36).slice(2, 8)}`;
-}
-
-function sanitizeName(name) {
-  return (name || "Player").trim() || "Player";
-}
-
-function makePlayer(id, name) {
-  return {
-    id,
-    name,
-    position: { x: 0, y: PLAYER_RADIUS, z: 0 },
-    velocity: { x: 0, z: 0 },
-    input: { x: 0, z: 0 },
-    heading: 0,
-    fallVelocityY: 0,
-    eliminated: false,
-    roundResult: null,
-  };
-}
-
-function makeRoom(password) {
-  return {
-    id: randomRoomId(),
-    password,
-    players: new Map(),
-    interval: null,
-    roundInProgress: false,
-  };
-}
-
-function resolveBallCollision(a, b) {
-  if (Math.abs(a.position.y - b.position.y) > PLAYER_RADIUS * 1.5) {
-    return;
-  }
-
-  const dx = b.position.x - a.position.x;
-  const dz = b.position.z - a.position.z;
-  const dist = Math.sqrt(dx * dx + dz * dz);
-  const minDist = PLAYER_RADIUS * 2;
-
-  if (dist >= minDist || dist < 0.0001) {
-    return;
-  }
-
-  const nx = dx / dist;
-  const nz = dz / dist;
-
-  const rvx = b.velocity.x - a.velocity.x;
-  const rvz = b.velocity.z - a.velocity.z;
-  const velAlongNormal = rvx * nx + rvz * nz;
-
-  if (velAlongNormal > 0) {
-    return;
-  }
-
-  const impulse =
-    (-(1 + RESTITUTION) * velAlongNormal) / (1 / PLAYER_MASS + 1 / PLAYER_MASS);
-
-  a.velocity.x -= (impulse / PLAYER_MASS) * nx;
-  a.velocity.z -= (impulse / PLAYER_MASS) * nz;
-  b.velocity.x += (impulse / PLAYER_MASS) * nx;
-  b.velocity.z += (impulse / PLAYER_MASS) * nz;
-
-  const overlap = (minDist - dist) / 2;
-  a.position.x -= overlap * nx;
-  a.position.z -= overlap * nz;
-  b.position.x += overlap * nx;
-  b.position.z += overlap * nz;
-}
-
-function placePlayerAtSpawnSlot(player, slotIndex) {
-  const isNorthSide = slotIndex % 2 === 0;
-
-  player.position = {
-    x: 0,
-    y: PLATE_SURFACE_Y,
-    z: isNorthSide ? SPAWN_DISTANCE : -SPAWN_DISTANCE,
-  };
-  player.heading = isNorthSide ? 0 : Math.PI;
-}
-
-function placePlayersForRound(room) {
-  const players = [...room.players.values()];
-
-  players.forEach((player, index) => {
-    placePlayerAtSpawnSlot(player, index);
-  });
-}
-
-function serializePlayers(playersMap) {
-  return [...playersMap.values()].map((player) => ({
-    id: player.id,
-    name: player.name,
-    position: player.position,
-    velocity: player.velocity,
-    heading: player.heading,
-    eliminated: player.eliminated,
-  }));
-}
+const {
+  TICK_RATE,
+  TICK_MS,
+  MAX_PLAYERS_PER_ROOM,
+  FALL_ELIMINATION_Y,
+  PLATE_SURFACE_Y,
+} = require("./constants");
+const {
+  clamp,
+  sanitizeName,
+  makePlayer,
+  makeRoom,
+  placePlayerAtSpawnSlot,
+  placePlayersForRound,
+  resetPlayerForRound,
+  stopPlayerHorizontalMovement,
+  serializePlayers,
+} = require("./player-utils");
+const {
+  resolveBallCollision,
+  applyFalling,
+  isPlayerOnPlate,
+  applyMovementInput,
+  capHorizontalSpeed,
+  applyHorizontalFriction,
+  advanceHorizontalPosition,
+} = require("./physics-utils");
 
 function createRoomService(io) {
   const rooms = new Map();
 
-  function resetPlayerForRound(player) {
-    player.velocity.x = 0;
-    player.velocity.z = 0;
-    player.input.x = 0;
-    player.input.z = 0;
-    player.fallVelocityY = 0;
-    player.position.y = PLATE_SURFACE_Y;
-    player.eliminated = false;
-    player.roundResult = null;
-  }
-
-  function stopPlayerHorizontalMovement(player) {
-    player.input.x = 0;
-    player.input.z = 0;
-    player.velocity.x = 0;
-    player.velocity.z = 0;
-  }
-
-  function applyFalling(player, dt) {
-    player.fallVelocityY -= GRAVITY * dt;
-    player.position.y += player.fallVelocityY * dt;
-  }
-
-  function isPlayerOnPlate(player) {
-    const distanceFromCenter = Math.hypot(player.position.x, player.position.z);
-    return distanceFromCenter <= PLATE_BOUNDARY_RADIUS;
-  }
-
-  function applyMovementInput(player, dt) {
-    player.heading += player.input.x * TURN_SPEED * dt;
-
-    const throttle = -player.input.z;
-    const forwardX = Math.sin(player.heading);
-    const forwardZ = -Math.cos(player.heading);
-
-    player.velocity.x += forwardX * throttle * ACCEL * dt;
-    player.velocity.z += forwardZ * throttle * ACCEL * dt;
-  }
-
-  function capHorizontalSpeed(player) {
-    const speed = Math.hypot(player.velocity.x, player.velocity.z);
-    if (speed <= MAX_SPEED) {
+  function removeRoom(roomId) {
+    const room = rooms.get(roomId);
+    if (!room) {
       return;
     }
 
-    const scale = MAX_SPEED / speed;
-    player.velocity.x *= scale;
-    player.velocity.z *= scale;
+    if (room.interval) {
+      clearInterval(room.interval);
+    }
+
+    rooms.delete(roomId);
   }
 
-  function applyHorizontalFriction(player) {
-    player.velocity.x *= FRICTION;
-    player.velocity.z *= FRICTION;
-  }
-
-  function advanceHorizontalPosition(player, dt) {
-    player.position.x += player.velocity.x * dt;
-    player.position.z += player.velocity.z * dt;
+  function broadcastRoomState(room) {
+    io.to(room.id).emit("roomState", {
+      roomId: room.id,
+      players: serializePlayers(room.players),
+    });
   }
 
   function notifyWaitingForOpponent(room) {
@@ -195,28 +59,24 @@ function createRoomService(io) {
     });
   }
 
-  function emitLoss(player) {
+  function emitRoundResult(player, result, message) {
     if (player.roundResult) {
       return;
     }
 
-    player.roundResult = "lost";
+    player.roundResult = result;
     io.to(player.id).emit("roundResult", {
-      result: "lost",
-      message: "You lose!",
+      result,
+      message,
     });
   }
 
-  function emitWin(player) {
-    if (player.roundResult) {
-      return;
-    }
+  function emitLoss(player) {
+    emitRoundResult(player, "lost", "You lose!");
+  }
 
-    player.roundResult = "won";
-    io.to(player.id).emit("roundResult", {
-      result: "won",
-      message: "You won!",
-    });
+  function emitWin(player) {
+    emitRoundResult(player, "won", "You won!");
   }
 
   function endRoundIfNeeded(room) {
@@ -260,24 +120,39 @@ function createRoomService(io) {
     });
   }
 
-  function removeRoom(roomId) {
-    const room = rooms.get(roomId);
-    if (!room) {
+  function tickPlayer(room, player, dt) {
+    if (player.eliminated) {
+      applyFalling(player, dt);
       return;
     }
 
-    if (room.interval) {
-      clearInterval(room.interval);
+    applyMovementInput(player, dt);
+    capHorizontalSpeed(player);
+    applyHorizontalFriction(player);
+    advanceHorizontalPosition(player, dt);
+
+    if (!isPlayerOnPlate(player)) {
+      stopPlayerHorizontalMovement(player);
+      applyFalling(player, dt);
+
+      if (player.position.y <= FALL_ELIMINATION_Y) {
+        eliminatePlayer(room, player);
+      }
+      return;
     }
 
-    rooms.delete(roomId);
+    player.position.y = PLATE_SURFACE_Y;
+    player.fallVelocityY = 0;
   }
 
-  function broadcastRoomState(room) {
-    io.to(room.id).emit("roomState", {
-      roomId: room.id,
-      players: serializePlayers(room.players),
-    });
+  function resolveActiveCollisions(players) {
+    const activePlayers = players.filter((player) => !player.eliminated);
+
+    for (let i = 0; i < activePlayers.length; i += 1) {
+      for (let j = i + 1; j < activePlayers.length; j += 1) {
+        resolveBallCollision(activePlayers[i], activePlayers[j]);
+      }
+    }
   }
 
   function tickRoom(room) {
@@ -289,40 +164,11 @@ function createRoomService(io) {
     const dt = TICK_MS / 1000;
 
     for (const player of players) {
-      if (player.eliminated) {
-        applyFalling(player, dt);
-        continue;
-      }
-
-      applyMovementInput(player, dt);
-      capHorizontalSpeed(player);
-      applyHorizontalFriction(player);
-      advanceHorizontalPosition(player, dt);
-
-      if (!isPlayerOnPlate(player)) {
-        stopPlayerHorizontalMovement(player);
-        applyFalling(player, dt);
-
-        if (player.position.y <= FALL_ELIMINATION_Y) {
-          eliminatePlayer(room, player);
-        }
-
-        continue;
-      }
-
-      player.position.y = PLATE_SURFACE_Y;
-      player.fallVelocityY = 0;
+      tickPlayer(room, player, dt);
     }
 
-    const activePlayers = players.filter((player) => !player.eliminated);
-    for (let i = 0; i < activePlayers.length; i += 1) {
-      for (let j = i + 1; j < activePlayers.length; j += 1) {
-        resolveBallCollision(activePlayers[i], activePlayers[j]);
-      }
-    }
-
+    resolveActiveCollisions(players);
     endRoundIfNeeded(room);
-
     broadcastRoomState(room);
   }
